リファクタ2
