+++ app/src/main/java/net/matsudamper/browser/BrowserToolBar.kt
@@ -4,10 +4,8 @@ import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectHorizontalDragGestures
 import androidx.compose.foundation.gestures.detectVerticalDragGestures
-import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.PaddingValues
@@ -19,16 +17,8 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
@@ -44,36 +34,22 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.autofill.ContentDataType
-import androidx.compose.ui.autofill.ContentType
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.luminance
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onSizeChanged
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.semantics.contentDataType
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.contentType
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.stateDescription
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
 
 @Composable
 internal fun BrowserToolBar(
@@ -207,29 +183,19 @@ internal fun BrowserToolbar(
 ) {
     var heightCache by remember { mutableIntStateOf(0) }
 
-    val isSystemDarkTheme = isSystemInDarkTheme()
-    val resolvedToolbarColor = toolbarColor ?: MaterialTheme.colorScheme.primaryContainer
-    val urlBarBackgroundColor: Color
-    val toolbarContentColor: Color
-    run {
-        val isBrightThemeColor = toolbarColor?.luminance()?.let { it >= 0.5f } ?: !isSystemDarkTheme
-        urlBarBackgroundColor = if (isBrightThemeColor) {
-            Color.Black
-        } else {
-            Color.LightGray
-        }
-        toolbarContentColor = if (isBrightThemeColor) Color.White else Color.Black
-    }
-
-    val colorSource = if (toolbarColor == null) "default" else "theme"
+    val toolbarColors = resolveBrowserToolbarColors(
+        toolbarColor = toolbarColor,
+        defaultToolbarColor = MaterialTheme.colorScheme.primaryContainer,
+        isSystemDarkTheme = isSystemInDarkTheme(),
+    )
 
     Surface(
-        color = resolvedToolbarColor,
-        contentColor = urlBarBackgroundColor,
+        color = toolbarColors.resolvedToolbarColor,
+        contentColor = toolbarColors.urlBarBackgroundColor,
         modifier = modifier
             .testTag(TEST_TAG_TOOLBAR)
             .semantics {
-                stateDescription = "toolbarColor|$colorSource|${resolvedToolbarColor.toArgbHex()}"
+                stateDescription = "toolbarColor|${toolbarColors.colorSource}|${toolbarColors.resolvedToolbarColor.toArgbHex()}"
             }
             .then(gestureState?.modifier ?: Modifier)
             .onSizeChanged {
@@ -247,8 +213,8 @@ internal fun BrowserToolbar(
                     .height(IntrinsicSize.Min)
                     .weight(1f)
                     .padding(4.dp),
-                contentColor = toolbarContentColor,
-                color = urlBarBackgroundColor,
+                contentColor = toolbarColors.toolbarContentColor,
+                color = toolbarColors.urlBarBackgroundColor,
                 shape = CircleShape,
             ) {
                 Row(
@@ -330,229 +296,7 @@ internal fun BrowserToolbar(
     }
 }
 
-@Composable
-private fun UrlTextInput(
-    value: String,
-    scrollEnabled: Boolean,
-    onValueChange: (String) -> Unit,
-    onSubmit: (String) -> Unit,
-    onFocusChanged: (Boolean) -> Unit,
-    textColor: Color,
-    enableSuggest: Boolean,
-    paddingValues: PaddingValues,
-    modifier: Modifier = Modifier,
-) {
-    val currentValue by rememberUpdatedState(value)
-    val currentOnValueChange by rememberUpdatedState(onValueChange)
-    val currentOnSubmit by rememberUpdatedState(onSubmit)
-    val currentOnFocusChanged by rememberUpdatedState(onFocusChanged)
-    val textColor by rememberUpdatedState(textColor)
-    val scrollEnabled by rememberUpdatedState(scrollEnabled)
-    val paddingValues by rememberUpdatedState(paddingValues)
-
-    // 一度AndroidViewを経由しないとBitwardenが認識しない
-    AndroidView(
-        modifier = modifier,
-        factory = { context ->
-            ComposeView(context).apply {
-                setContent {
-                    BasicTextField(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .horizontalScroll(
-                                state = rememberScrollState(),
-                                enabled = scrollEnabled,
-                            )
-                            .padding(paddingValues)
-                            .then(
-                                if (enableSuggest) {
-                                    Modifier.testTag("url_bar")
-                                } else {
-                                    Modifier
-                                }
-                            )
-
-                            .onFocusChanged { currentOnFocusChanged(it.hasFocus) }
-                            .semantics {
-                                if (enableSuggest) {
-                                    contentDescription = "Address bar"
-                                    contentType = ContentType("url")
-                                    contentDataType = ContentDataType.Text
-                                }
-                            },
-                        value = currentValue,
-                        onValueChange = { currentOnValueChange(it) },
-                        singleLine = true,
-                        textStyle = MaterialTheme.typography.bodyLarge
-                            .merge(
-                                color = textColor,
-                                textAlign = TextAlign.Start,
-                            ),
-                        cursorBrush = SolidColor(MaterialTheme.colorScheme.primary),
-                        keyboardOptions = KeyboardOptions.Default.copy(
-                            imeAction = ImeAction.Go,
-                            keyboardType = KeyboardType.Uri,
-                            autoCorrectEnabled = false,
-                        ),
-                        keyboardActions = KeyboardActions(
-                            onGo = { currentOnSubmit(currentValue) },
-                            onDone = { currentOnSubmit(currentValue) },
-                            onSearch = { currentOnSubmit(currentValue) },
-                        ),
-                    )
-                }
-            }
-        },
-    )
-}
-
 internal const val TEST_TAG_TOOLBAR = "browser_toolbar"
-
-private fun Color.toArgbHex(): String {
-    val raw = toArgb().toUInt().toString(16).padStart(8, '0')
-    return "#${raw.uppercase()}"
-}
-
-@Composable
-private fun ToolbarMenu(
-    visibleMenu: Boolean,
-    onDismissRequest: () -> Unit,
-    onRefresh: () -> Unit,
-    onHome: () -> Unit,
-    onForward: () -> Unit,
-    canGoForward: Boolean,
-    isPcMode: Boolean,
-    onPcModeToggle: () -> Unit,
-    showInstallExtensionItem: Boolean,
-    onInstallExtension: () -> Unit,
-    onTranslatePage: () -> Unit,
-    onShare: () -> Unit,
-    onFindInPage: () -> Unit,
-    onOpenSettings: () -> Unit,
-) {
-    DropdownMenu(
-        expanded = visibleMenu,
-        onDismissRequest = { onDismissRequest() }
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 8.dp, vertical = 4.dp),
-            horizontalArrangement = Arrangement.SpaceEvenly,
-        ) {
-            Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                IconButton(
-                    onClick = {
-                        onDismissRequest()
-                        onForward()
-                    },
-                    enabled = canGoForward,
-                ) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_arrow_forward_24dp),
-                        contentDescription = "進む",
-                    )
-                }
-                Text(
-                    text = "進む",
-                    style = MaterialTheme.typography.labelSmall,
-                )
-            }
-            Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                IconButton(
-                    onClick = {
-                        onDismissRequest()
-                        onHome()
-                    }
-                ) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_home_24dp),
-                        contentDescription = "ホーム",
-                    )
-                }
-                Text(
-                    text = "ホーム",
-                    style = MaterialTheme.typography.labelSmall,
-                )
-            }
-            Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                IconButton(
-                    onClick = {
-                        onDismissRequest()
-                        onRefresh()
-                    }
-                ) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_refresh_24dp),
-                        contentDescription = "更新",
-                    )
-                }
-                Text(
-                    text = "更新",
-                    style = MaterialTheme.typography.labelSmall,
-                )
-            }
-        }
-        HorizontalDivider()
-        DropdownMenuItem(
-            text = { Text(text = "PCページ") },
-            leadingIcon = {
-                Checkbox(
-                    checked = isPcMode,
-                    onCheckedChange = null,
-                )
-            },
-            onClick = { onPcModeToggle() },
-        )
-        if (showInstallExtensionItem) {
-            DropdownMenuItem(
-                text = {
-                    Text(text = "拡張機能をインストール")
-                },
-                onClick = {
-                    onDismissRequest()
-                    onInstallExtension()
-                },
-            )
-        }
-        DropdownMenuItem(
-            text = {
-                Text(text = "翻訳")
-            },
-            onClick = {
-                onDismissRequest()
-                onTranslatePage()
-            },
-        )
-        DropdownMenuItem(
-            text = {
-                Text(text = "共有")
-            },
-            onClick = {
-                onDismissRequest()
-                onShare()
-            },
-        )
-        DropdownMenuItem(
-            text = {
-                Text(text = "ページ内検索")
-            },
-            onClick = {
-                onDismissRequest()
-                onFindInPage()
-            },
-        )
-        DropdownMenuItem(
-            text = {
-                Text(text = "設定")
-            },
-            onClick = {
-                onDismissRequest()
-                onOpenSettings()
-            },
-        )
-    }
-}
 
 @Preview(name = "Light")
 @Preview(name = "Dark", uiMode = android.content.res.Configuration.UI_MODE_NIGHT_YES)
