+++ app/src/main/java/net/matsudamper/browser/BrowserToolbarMenu.kt
@@ -1,0 +1,161 @@
+package net.matsudamper.browser
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun ToolbarMenu(
+    visibleMenu: Boolean,
+    onDismissRequest: () -> Unit,
+    onRefresh: () -> Unit,
+    onHome: () -> Unit,
+    onForward: () -> Unit,
+    canGoForward: Boolean,
+    isPcMode: Boolean,
+    onPcModeToggle: () -> Unit,
+    showInstallExtensionItem: Boolean,
+    onInstallExtension: () -> Unit,
+    onTranslatePage: () -> Unit,
+    onShare: () -> Unit,
+    onFindInPage: () -> Unit,
+    onOpenSettings: () -> Unit,
+) {
+    DropdownMenu(
+        expanded = visibleMenu,
+        onDismissRequest = { onDismissRequest() }
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 8.dp, vertical = 4.dp),
+            horizontalArrangement = Arrangement.SpaceEvenly,
+        ) {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                IconButton(
+                    onClick = {
+                        onDismissRequest()
+                        onForward()
+                    },
+                    enabled = canGoForward,
+                ) {
+                    Icon(
+                        painter = painterResource(R.drawable.ic_arrow_forward_24dp),
+                        contentDescription = "進む",
+                    )
+                }
+                Text(
+                    text = "進む",
+                    style = MaterialTheme.typography.labelSmall,
+                )
+            }
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                IconButton(
+                    onClick = {
+                        onDismissRequest()
+                        onHome()
+                    }
+                ) {
+                    Icon(
+                        painter = painterResource(R.drawable.ic_home_24dp),
+                        contentDescription = "ホーム",
+                    )
+                }
+                Text(
+                    text = "ホーム",
+                    style = MaterialTheme.typography.labelSmall,
+                )
+            }
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                IconButton(
+                    onClick = {
+                        onDismissRequest()
+                        onRefresh()
+                    }
+                ) {
+                    Icon(
+                        painter = painterResource(R.drawable.ic_refresh_24dp),
+                        contentDescription = "更新",
+                    )
+                }
+                Text(
+                    text = "更新",
+                    style = MaterialTheme.typography.labelSmall,
+                )
+            }
+        }
+        HorizontalDivider()
+        DropdownMenuItem(
+            text = { Text(text = "PCページ") },
+            leadingIcon = {
+                Checkbox(
+                    checked = isPcMode,
+                    onCheckedChange = null,
+                )
+            },
+            onClick = { onPcModeToggle() },
+        )
+        if (showInstallExtensionItem) {
+            DropdownMenuItem(
+                text = {
+                    Text(text = "拡張機能をインストール")
+                },
+                onClick = {
+                    onDismissRequest()
+                    onInstallExtension()
+                },
+            )
+        }
+        DropdownMenuItem(
+            text = {
+                Text(text = "翻訳")
+            },
+            onClick = {
+                onDismissRequest()
+                onTranslatePage()
+            },
+        )
+        DropdownMenuItem(
+            text = {
+                Text(text = "共有")
+            },
+            onClick = {
+                onDismissRequest()
+                onShare()
+            },
+        )
+        DropdownMenuItem(
+            text = {
+                Text(text = "ページ内検索")
+            },
+            onClick = {
+                onDismissRequest()
+                onFindInPage()
+            },
+        )
+        DropdownMenuItem(
+            text = {
+                Text(text = "設定")
+            },
+            onClick = {
+                onDismissRequest()
+                onOpenSettings()
+            },
+        )
+    }
+}
