+++ app/src/main/java/net/matsudamper/browser/GeckoBrowserTab.kt
@@ -10,38 +10,23 @@ import android.view.ViewGroup
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
 import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.ListItem
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -68,19 +53,16 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.graphics.toColorInt
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import kotlinx.coroutines.flow.collectLatest
 import net.matsudamper.browser.data.TranslationProvider
+import net.matsudamper.browser.media.GeckoMediaSessionDelegate
 import net.matsudamper.browser.media.MediaWebExtension
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoSession
 import org.mozilla.geckoview.GeckoView
-import org.mozilla.geckoview.PanZoomController
 
 @Composable
 @OptIn(ExperimentalLayoutApi::class)
@@ -199,24 +181,24 @@ internal fun GeckoBrowserTab(
     }
 
     DisposableEffect(session, state, browserTab) {
-        val permissionDelegate = state.createPermissionDelegate { uri ->
-            currentOnDesktopNotificationPermissionRequest(uri)
-        }
-        val navigationDelegate = state.createNavigationDelegate { uri ->
-            currentOnOpenNewSessionRequest(uri)
-        }
-        val contentDelegate = state.createContentDelegate(
-            onClose = { currentOnCloseTab?.invoke() },
+        val delegates = createGeckoSessionDelegateBundle(
+            callbacks = state,
+            onDesktopNotificationPermissionRequest = { uri ->
+                currentOnDesktopNotificationPermissionRequest(uri)
+            },
+            onOpenNewSessionRequest = { uri ->
+                currentOnOpenNewSessionRequest(uri)
+            },
+            onCloseRequest = { currentOnCloseTab?.invoke() },
         )
-        val progressDelegate = state.createProgressDelegate()
-        val translationsDelegate = state.createTranslationsDelegate()
         val promptDelegate = dialogState.createPromptDelegate()
 
-        session.permissionDelegate = permissionDelegate
-        session.navigationDelegate = navigationDelegate
-        session.contentDelegate = contentDelegate
-        session.progressDelegate = progressDelegate
-        session.translationsSessionDelegate = translationsDelegate
+        session.permissionDelegate = delegates.permissionDelegate
+        session.navigationDelegate = delegates.navigationDelegate
+        session.contentDelegate = delegates.contentDelegate
+        session.progressDelegate = delegates.progressDelegate
+        session.translationsSessionDelegate = delegates.translationsDelegate
+        session.scrollDelegate = delegates.scrollDelegate
         session.promptDelegate = promptDelegate
 
         browserSessionController.restoreSession(browserTab)
@@ -227,27 +209,19 @@ internal fun GeckoBrowserTab(
             session.contentDelegate = null
             session.progressDelegate = null
             session.translationsSessionDelegate = null
+            session.scrollDelegate = null
             session.promptDelegate = null
         }
     }
 
     // メディアセッションデリゲートは不安定なラムダキーの影響を受けないよう独立して管理
-    DisposableEffect(session, state, mediaWebExtension) {
-        val mediaSessionDelegate = state.createMediaSessionDelegate(mediaWebExtension)
+    DisposableEffect(session, mediaWebExtension) {
+        val mediaSessionDelegate = GeckoMediaSessionDelegate(mediaWebExtension)
         session.mediaSessionDelegate = mediaSessionDelegate
         onDispose {
             if (session.mediaSessionDelegate === mediaSessionDelegate) {
                 session.mediaSessionDelegate = null
             }
-        }
-    }
-
-    // Attach scroll delegate
-    DisposableEffect(session, state) {
-        val scrollDelegate = state.createScrollDelegate()
-        session.scrollDelegate = scrollDelegate
-        onDispose {
-            if (session.scrollDelegate === scrollDelegate) session.scrollDelegate = null
         }
     }
 
@@ -383,448 +357,22 @@ internal fun GeckoBrowserTab(
                 }
             )
 
-            state.pageLoadError?.let { pageLoadError ->
-                PageLoadErrorOverlay(
-                    pageLoadError = pageLoadError,
-                    onRetry = state::retryPageLoad,
-                )
-            }
-
-            // URLバーフォーカス中にサジェストを表示
-            if (
-                !state.showFindInPage &&
-                state.isUrlInputFocused &&
-                (historySuggestions.isNotEmpty() || state.currentPageUrl.isNotBlank())
-            ) {
-                HistorySuggestionList(
-                    currentPageUrl = state.currentPageUrl,
-                    suggestions = historySuggestions,
-                    onSuggestionClick = { entry ->
-                        state.onUrlSubmit(entry.url)
-                        closeUrlInput(false)
-                    },
-                    onCopyCurrentUrl = state::copyCurrentPageUrl,
-                    onRestoreCurrentUrl = state::restoreCurrentPageUrlToInput,
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .testTag(TEST_TAG_HISTORY_SUGGESTION_LIST)
-                        .background(MaterialTheme.colorScheme.surface),
-                )
-            }
-        }
-
-        // Image context menu dialog
-        state.imageContextMenuUrl?.let { imageUrl ->
-            AlertDialog(
-                onDismissRequest = { state.imageContextMenuUrl = null },
-                title = { Text(text = "画像") },
-                text = { Text(text = "この画像をダウンロードしますか？") },
-                confirmButton = {
-                    TextButton(onClick = { state.downloadImage(imageUrl) }) {
-                        Text(text = "ダウンロード")
-                    }
+            BrowserTabOverlayLayer(
+                state = state,
+                historySuggestions = historySuggestions,
+                onSuggestionClick = { entry ->
+                    state.onUrlSubmit(entry.url)
+                    closeUrlInput(false)
                 },
-                dismissButton = {
-                    TextButton(onClick = { state.imageContextMenuUrl = null }) {
-                        Text(text = "キャンセル")
-                    }
-                },
+                modifier = Modifier.fillMaxSize(),
             )
         }
-
-        // Link context menu dialog
-        state.linkContextMenuUrl?.let { linkUrl ->
-            AlertDialog(
-                onDismissRequest = { state.linkContextMenuUrl = null },
-                title = { Text(text = "リンク") },
-                text = {
-                    Text(
-                        text = linkUrl,
-                        maxLines = 3,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                },
-                confirmButton = {
-                    TextButton(onClick = { state.copyLinkUrl(linkUrl) }) {
-                        Text("URLをコピー")
-                    }
-                },
-                dismissButton = {
-                    Column {
-                        if (enableTabUi) {
-                            TextButton(
-                                onClick = {
-                                    currentOnOpenNewSessionRequest(linkUrl)
-                                    state.linkContextMenuUrl = null
-                                },
-                            ) {
-                                Text("新しいタブで開く")
-                            }
-                        } else {
-                            TextButton(
-                                onClick = {
-                                    state.onUrlSubmit(linkUrl)
-                                    state.linkContextMenuUrl = null
-                                },
-                            ) {
-                                Text("開く")
-                            }
-                        }
-                        TextButton(onClick = { state.linkContextMenuUrl = null }) {
-                            Text("キャンセル")
-                        }
-                    }
-                },
-            )
-        }
-
-        // ファイルダウンロード確認ダイアログ
-        state.pendingDownloadResponse?.let { response ->
-            AlertDialog(
-                onDismissRequest = state::dismissPendingDownload,
-                title = { Text("ダウンロード") },
-                text = {
-                    Text(
-                        text = response.uri,
-                        maxLines = 4,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                },
-                confirmButton = {
-                    TextButton(onClick = state::confirmPendingDownload) {
-                        Text("ダウンロード")
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = state::dismissPendingDownload) {
-                        Text("キャンセル")
-                    }
-                },
-            )
-        }
-
-        // Alert prompt dialog
-        dialogState.pendingAlertPrompt?.let { prompt ->
-            AlertDialog(
-                onDismissRequest = dialogState::dismissAlertPrompt,
-                text = { Text(prompt.message ?: "") },
-                confirmButton = {
-                    TextButton(onClick = dialogState::dismissAlertPrompt) {
-                        Text("OK")
-                    }
-                },
-            )
-        }
-
-        // Button prompt dialog (window.confirm())
-        dialogState.pendingButtonPrompt?.let { prompt ->
-            AlertDialog(
-                onDismissRequest = dialogState::dismissButtonPrompt,
-                text = { Text(prompt.message ?: "") },
-                confirmButton = {
-                    TextButton(onClick = { dialogState.confirmButtonPrompt(true) }) {
-                        Text("OK")
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = { dialogState.confirmButtonPrompt(false) }) {
-                        Text("キャンセル")
-                    }
-                },
-            )
-        }
-
-        // Text prompt dialog (window.prompt())
-        dialogState.pendingTextPrompt?.let { prompt ->
-            var textValue by remember(prompt) { mutableStateOf(prompt.defaultValue ?: "") }
-            AlertDialog(
-                onDismissRequest = dialogState::dismissTextPrompt,
-                title = prompt.message?.takeIf { it.isNotEmpty() }?.let { { Text(it) } },
-                text = {
-                    OutlinedTextField(
-                        value = textValue,
-                        onValueChange = { textValue = it },
-                        singleLine = true,
-                    )
-                },
-                confirmButton = {
-                    TextButton(onClick = { dialogState.confirmTextPrompt(textValue) }) {
-                        Text("OK")
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = dialogState::dismissTextPrompt) {
-                        Text("キャンセル")
-                    }
-                },
-            )
-        }
-
-        // Choice prompt dialog (<select> elements)
-        dialogState.pendingChoicePrompt?.let { prompt ->
-            ChoicePromptDialog(
-                prompt = prompt,
-                onDismiss = dialogState::dismissChoicePrompt,
-                onConfirmSingle = dialogState::confirmChoicePromptSingle,
-                onConfirmMultiple = dialogState::confirmChoicePromptMultiple,
-            )
-        }
-
-        // Color prompt dialog (<input type="color">)
-        dialogState.pendingColorPrompt?.let { prompt ->
-            var colorText by remember(prompt) { mutableStateOf(prompt.defaultValue ?: "#000000") }
-            val parsedColor = remember(colorText) {
-                runCatching {
-                    androidx.compose.ui.graphics.Color(colorText.toColorInt())
-                }.getOrNull()
-            }
-            AlertDialog(
-                onDismissRequest = dialogState::dismissColorPrompt,
-                title = { Text("色を選択") },
-                text = {
-                    Column {
-                        if (parsedColor != null) {
-                            Box(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(48.dp)
-                                    .background(parsedColor),
-                            )
-                            Spacer(modifier = Modifier.height(8.dp))
-                        }
-                        OutlinedTextField(
-                            value = colorText,
-                            onValueChange = { colorText = it },
-                            label = { Text("#RRGGBB") },
-                            singleLine = true,
-                        )
-                    }
-                },
-                confirmButton = {
-                    TextButton(
-                        onClick = { dialogState.confirmColorPrompt(colorText) },
-                        enabled = parsedColor != null,
-                    ) {
-                        Text("OK")
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = dialogState::dismissColorPrompt) {
-                        Text("キャンセル")
-                    }
-                },
-            )
-        }
-
-        // DateTime prompt dialog (<input type="date/time/...">)
-        dialogState.pendingDateTimePrompt?.let { prompt ->
-            var dateTimeText by remember(prompt) { mutableStateOf(prompt.defaultValue ?: "") }
-            val (title, hint) = when (prompt.type) {
-                GeckoSession.PromptDelegate.DateTimePrompt.Type.DATE ->
-                    "日付を選択" to "YYYY-MM-DD"
-
-                GeckoSession.PromptDelegate.DateTimePrompt.Type.TIME ->
-                    "時刻を選択" to "HH:MM"
-
-                GeckoSession.PromptDelegate.DateTimePrompt.Type.MONTH ->
-                    "年月を選択" to "YYYY-MM"
-
-                GeckoSession.PromptDelegate.DateTimePrompt.Type.WEEK ->
-                    "週を選択" to "YYYY-Www"
-
-                GeckoSession.PromptDelegate.DateTimePrompt.Type.DATETIME_LOCAL ->
-                    "日時を選択" to "YYYY-MM-DDTHH:MM"
-
-                else -> "値を入力" to ""
-            }
-            AlertDialog(
-                onDismissRequest = dialogState::dismissDateTimePrompt,
-                title = { Text(title) },
-                text = {
-                    OutlinedTextField(
-                        value = dateTimeText,
-                        onValueChange = { dateTimeText = it },
-                        label = { Text(hint) },
-                        singleLine = true,
-                    )
-                },
-                confirmButton = {
-                    TextButton(onClick = { dialogState.confirmDateTimePrompt(dateTimeText) }) {
-                        Text("OK")
-                    }
-                },
-                dismissButton = {
-                    TextButton(onClick = dialogState::dismissDateTimePrompt) {
-                        Text("キャンセル")
-                    }
-                },
-            )
-        }
-    }
-}
-
-@Composable
-private fun CustomTabToolbar(
-    title: String,
-    url: String,
-    onClose: () -> Unit,
-    onRefresh: () -> Unit,
-    onShare: () -> Unit,
-    onOpenInBrowser: (() -> Unit)?,
-    toolbarColor: Color?,
-) {
-    var menuExpanded by remember { mutableStateOf(false) }
-    val resolvedToolbarColor = toolbarColor ?: MaterialTheme.colorScheme.primaryContainer
-    val toolbarContentColor = if (resolvedToolbarColor.luminance() >= 0.5f) Color.Black else Color.White
-    val toolbarSecondaryContentColor = toolbarContentColor.copy(alpha = 0.72f)
-
-    Surface(
-        color = resolvedToolbarColor,
-        contentColor = toolbarContentColor,
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 4.dp, vertical = 6.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            IconButton(onClick = onClose) {
-                Icon(
-                    painter = painterResource(R.drawable.close_24dp),
-                    contentDescription = "閉じる",
-                )
-            }
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(horizontal = 4.dp),
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleSmall,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-                Text(
-                    text = url,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = toolbarSecondaryContentColor,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-            Box {
-                IconButton(onClick = { menuExpanded = true }) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_more_vert_24dp),
-                        contentDescription = "メニュー",
-                    )
-                }
-                DropdownMenu(
-                    expanded = menuExpanded,
-                    onDismissRequest = { menuExpanded = false },
-                ) {
-                    onOpenInBrowser?.let { openInBrowser ->
-                        DropdownMenuItem(
-                            text = { Text("ブラウザで開く") },
-                            onClick = {
-                                menuExpanded = false
-                                openInBrowser()
-                            },
-                        )
-                    }
-                    DropdownMenuItem(
-                        text = { Text("共有") },
-                        onClick = {
-                            menuExpanded = false
-                            onShare()
-                        },
-                    )
-                    DropdownMenuItem(
-                        text = { Text("更新") },
-                        onClick = {
-                            menuExpanded = false
-                            onRefresh()
-                        },
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ChoicePromptDialog(
-    prompt: GeckoSession.PromptDelegate.ChoicePrompt,
-    onDismiss: () -> Unit,
-    onConfirmSingle: (GeckoSession.PromptDelegate.ChoicePrompt.Choice) -> Unit,
-    onConfirmMultiple: (Array<GeckoSession.PromptDelegate.ChoicePrompt.Choice>) -> Unit,
-) {
-    val isMultiple = prompt.type == GeckoSession.PromptDelegate.ChoicePrompt.Type.MULTIPLE
-    val flatChoices = remember(prompt) { flattenChoices(prompt.choices) }
-    val selectedIds = remember(prompt) {
-        mutableStateOf(flatChoices.filter { it.selected }.map { it.id }.toSet())
-    }
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        text = {
-            LazyColumn {
-                items(flatChoices) { choice ->
-                    if (choice.separator) {
-                        HorizontalDivider()
-                    } else {
-                        val isSelected = choice.id in selectedIds.value
-                        ListItem(
-                            headlineContent = { Text(choice.label) },
-                            leadingContent = {
-                                if (isMultiple) {
-                                    Checkbox(checked = isSelected, onCheckedChange = null)
-                                } else {
-                                    RadioButton(selected = isSelected, onClick = null)
-                                }
-                            },
-                            modifier = Modifier.clickable(enabled = !choice.disabled) {
-                                if (isMultiple) {
-                                    selectedIds.value = if (isSelected) {
-                                        selectedIds.value - choice.id
-                                    } else {
-                                        selectedIds.value + choice.id
-                                    }
-                                } else {
-                                    onConfirmSingle(choice)
-                                }
-                            },
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            if (isMultiple) {
-                TextButton(onClick = {
-                    val selected = flatChoices
-                        .filter { it.id in selectedIds.value }
-                        .toTypedArray()
-                    onConfirmMultiple(selected)
-                }) {
-                    Text("OK")
-                }
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("キャンセル")
-            }
-        },
-    )
-}
-
-private fun flattenChoices(
-    choices: Array<GeckoSession.PromptDelegate.ChoicePrompt.Choice>,
-): List<GeckoSession.PromptDelegate.ChoicePrompt.Choice> {
-    return choices.flatMap { choice ->
-        if (choice.items != null) choice.items!!.toList() else listOf(choice)
+        BrowserTabDialogLayer(
+            state = state,
+            dialogState = dialogState,
+            enableTabUi = enableTabUi,
+            onOpenNewSessionRequest = currentOnOpenNewSessionRequest,
+        )
     }
 }
 private const val URL_BAR_IME_HIDE_GRACE_MS = 700L
